Features
The facemask detector is accurate and uses no dataset of altered masked photos. It is computationally effective since MobileNetV2 architecture is utilized.
With  the event Covid-19 outbreak, this technology can therefore be employed in real-time applications that demand face-mask detection for security reasons. To ensure that public safety regulations are obeyed, this project can be connected with embedded technologies for application at airports, train stations, offices, schools, and public spaces.